Change type of RoutePriceMetric id to long

+++ flycierge/src/main/java/com/mgieroba/flycierge/model/RoutePriceMetric.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 @Builder
 @Getter
 public class RoutePriceMetric {
-    private double routePriceMetricId;
+    private long routePriceMetricId;
     private String origin;
     private String destination;
     private LocalDate departureDate;
